Sync TensorFlowSharp operations with native API (#91)
Just noticed that Operations.g.cs is not synchronized with native API provided in the nuget package.

Ran OpGenerator against the libtensorflow.dll from the latest (v1.2.2) nuget package. It might be a breaking change because some methods are no longer available, but for those who use the bindings via nuget it should not be a problem, because underlying native API doesn't provide those.

Added C# unit tests project and added coverage on 'BitwiseAnd' operation which was missing in the TensorFlowSharp, but available in the native API targeted OpGenerator to x64
added xunit runner for visual studio 2017. Now it allows running both F# and C# tests from VS 2017